separate model and block (#1154)
